rename and remove embarrassing console.log
