+++ widgets/src/lib/TableMetaCell.svelte
